Move function to question-submission.ts to avoid circular dependency
