+++ apps/prairielearn/src/lib/issues.ts
@@ -1,13 +1,9 @@
 import * as async from 'async';
-import { Request, Response } from 'express';
-import _ from 'lodash';
 
-import { HttpStatusError } from '@prairielearn/error';
 import * as sqldb from '@prairielearn/postgres';
 import { recursivelyTruncateStrings } from '@prairielearn/sanitize';
 
 import { Variant } from './db-types.js';
-import { validateVariantAgainstQuestion } from './question-submission.js';
 
 interface IssueForErrorData {
   variantId: string;
@@ -102,43 +98,4 @@ export async function writeCourseIssues(
       authnUserId: authn_user_id,
     });
   });
-}
-
-export async function reportIssueFromForm(
-  context: 'student' | 'instructor',
-  req: Request,
-  res: Response,
-): Promise<string> {
-  if (context === 'student' && !res.locals.assessment.allow_issue_reporting) {
-    throw new HttpStatusError(403, 'Issue reporting not permitted for this assessment');
-  }
-  const description = req.body.description;
-  if (typeof description !== 'string' || description.length === 0) {
-    throw new HttpStatusError(400, 'A description of the issue must be provided');
-  }
-
-  const variantId = (
-    await validateVariantAgainstQuestion(
-      req.body.__variant_id,
-      res.locals.question.id,
-      context === 'student' ? res.locals.instance_question?.id : null,
-    )
-  ).id;
-  await insertIssue({
-    variantId,
-    studentMessage: description,
-    instructorMessage: `${context}-reported issue`,
-    manuallyReported: true,
-    courseCaused: true,
-    courseData: _.pick(res.locals, [
-      'variant',
-      'question',
-      'course_instance',
-      'course',
-      ...(context === 'student' ? ['instance_question', 'assessment_instance', 'assessment'] : []),
-    ]),
-    systemData: {},
-    authnUserId: res.locals.authn_user.user_id,
-  });
-  return variantId;
 }
