+++ apps/prairielearn/src/lib/question-submission.ts
@@ -1,13 +1,14 @@
 import { type Request, type Response } from 'express';
 import _ from 'lodash';
 
-import * as error from '@prairielearn/error';
+import { HttpStatusError } from '@prairielearn/error';
 
 import { selectVariantById } from '../models/variant.js';
 
 import { type Variant } from './db-types.js';
 import { saveAndGradeSubmission, saveSubmission } from './grading.js';
 import { idsEqual } from './id.js';
+import { insertIssue } from './issues.js';
 
 export async function validateVariantAgainstQuestion(
   unsafe_variant_id: string,
@@ -16,7 +17,7 @@ export async function validateVariantAgainstQuestion(
 ): Promise<Variant> {
   const variant = await selectVariantById(unsafe_variant_id);
   if (variant == null || !idsEqual(variant.question_id, question_id)) {
-    throw new error.HttpStatusError(
+    throw new HttpStatusError(
       400,
       `Client-provided variant ID ${unsafe_variant_id} is not valid for question ID ${question_id}.`,
     );
@@ -25,7 +26,7 @@ export async function validateVariantAgainstQuestion(
     instance_question_id != null &&
     (!variant.instance_question_id || !idsEqual(variant.instance_question_id, instance_question_id))
   ) {
-    throw new error.HttpStatusError(
+    throw new HttpStatusError(
       400,
       `Client-provided variant ID ${unsafe_variant_id} is not valid for instance question ID ${instance_question_id}.`,
     );
@@ -44,13 +45,13 @@ export async function processSubmission(
     submitted_answer = _.omit(req.body, ['__action', '__csrf_token', '__variant_id']);
   } else {
     if (!req.body.postData) {
-      throw new error.HttpStatusError(400, 'No postData');
+      throw new HttpStatusError(400, 'No postData');
     }
     let postData;
     try {
       postData = JSON.parse(req.body.postData);
     } catch (e) {
-      throw new error.HttpStatusError(400, 'JSON parse failed on body.postData');
+      throw new HttpStatusError(400, 'JSON parse failed on body.postData');
     }
     variant_id = postData.variant ? postData.variant.id : null;
     submitted_answer = postData.submittedAnswer;
@@ -82,7 +83,7 @@ export async function processSubmission(
   // force-breaks variants, as we could be in a case where the variant wasn't
   // broken when the user loaded the page but it is broken when they submit.
   if (variant.broken_at) {
-    throw new error.HttpStatusError(403, 'Cannot submit to a broken variant');
+    throw new HttpStatusError(403, 'Cannot submit to a broken variant');
   }
 
   if (req.body.__action === 'grade') {
@@ -99,6 +100,45 @@ export async function processSubmission(
     await saveSubmission(submission, variant, res.locals.question, res.locals.course);
     return submission.variant_id;
   } else {
-    throw new error.HttpStatusError(400, `unknown __action: ${req.body.__action}`);
+    throw new HttpStatusError(400, `unknown __action: ${req.body.__action}`);
   }
+}
+
+export async function reportIssueFromForm(
+  context: 'student' | 'instructor',
+  req: Request,
+  res: Response,
+): Promise<string> {
+  if (context === 'student' && !res.locals.assessment.allow_issue_reporting) {
+    throw new HttpStatusError(403, 'Issue reporting not permitted for this assessment');
+  }
+  const description = req.body.description;
+  if (typeof description !== 'string' || description.length === 0) {
+    throw new HttpStatusError(400, 'A description of the issue must be provided');
+  }
+
+  const variantId = (
+    await validateVariantAgainstQuestion(
+      req.body.__variant_id,
+      res.locals.question.id,
+      context === 'student' ? res.locals.instance_question?.id : null,
+    )
+  ).id;
+  await insertIssue({
+    variantId,
+    studentMessage: description,
+    instructorMessage: `${context}-reported issue`,
+    manuallyReported: true,
+    courseCaused: true,
+    courseData: _.pick(res.locals, [
+      'variant',
+      'question',
+      'course_instance',
+      'course',
+      ...(context === 'student' ? ['instance_question', 'assessment_instance', 'assessment'] : []),
+    ]),
+    systemData: {},
+    authnUserId: res.locals.authn_user.user_id,
+  });
+  return variantId;
 }
